tabular_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
aiohttp=3.7.1=pypi_0
aiohttp-cors=0.7.0=pypi_0
aioredis=1.3.1=pypi_0
anyio=2.2.0=pypi_0
argon2-cffi=20.1.0=pypi_0
async-timeout=3.0.1=pypi_0
async_generator=1.10=py_0
attrs=20.2.0=pypi_0
babel=2.9.0=pypi_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.3=pypi_0
blas=1.0=mkl
bleach=3.2.1=py_0
blessings=1.7=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.1.1=pypi_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.14.5=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=pypi_0
colorama=0.4.4=pypi_0
colorful=0.5.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cvxopt=1.2.5=py38hf48cded_2
cvxpy=1.1.7=py38h950e882_0
cvxpy-base=1.1.7=py38hc5bc63f_0
cycler=0.10.0=py38_0
cython=0.29.27=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
deprecation=2.1.0=pypi_0
docker-pycreds=0.4.0=pypi_0
dsdp=5.8=hd9d9efa_1203
easyprocess=0.3=py38h32f6830_2
ecos=2.0.7.post1=py38h0b5ebd8_3
entrypoints=0.3=py38_0
et-xmlfile=1.0.1=pypi_0
expat=2.2.10=he6710b0_2
fasteners=0.17.3=pypi_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.8=nompi_h7f3a6c3_1111
filelock=3.0.12=pypi_0
flake8=3.8.4=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.3=h5ab3b9f_0
future=0.18.2=py38h578d9bd_2
gitdb=4.0.9=pypi_0
gitpython=3.1.26=pypi_0
glfw=2.5.0=pypi_0
glib=2.66.1=h92f7085_0
glpk=4.65=he80fd80_1002
gmp=6.2.1=h58526e2_0
gnutls=3.6.15=he1e5248_0
google=3.0.0=pypi_0
google-api-core=1.23.0=pypi_0
google-auth=1.22.1=pypi_0
googleapis-common-protos=1.52.0=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.33.1=pypi_0
gsl=2.6=hf94e986_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
gym=0.21.0=pypi_0
hiredis=1.1.0=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pypi_0
imageio=2.16.0=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=py38_0
jdcal=1.4.1=pypi_0
jedi=0.17.2=py38_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pypi_0
jsonschema=3.2.0=py_2
jupyter-lsp=1.1.4=pypi_0
jupyter-packaging=0.9.2=pypi_0
jupyter-server=1.6.2=pypi_0
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py38_0
jupyterlab=3.0.14=pypi_0
jupyterlab-lsp=3.5.0=pypi_0
jupyterlab-server=2.4.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.2.0=py38hfd86e86_0
krb5=1.17.1=hfafb76e_3
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=20_mkl
libcblas=3.8.0=20_mkl
libedit=3.1.20191231=he28a2e2_2
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
libidn2=2.3.2=h7f8727e_0
liblapack=3.8.0=20_mkl
libpng=1.6.37=hbc83047_0
libpq=12.3=h5513abc_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=he19cac6_1
lz4-c=1.9.2=heb0550a_3
markupsafe=1.1.1=py38h7b6447c_0
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py38h817c723_0
mccabe=0.6.1=py38_1
metis=5.1.0=h58526e2_1006
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
msgpack=1.0.0=pypi_0
multidict=5.0.0=pypi_0
nbclassic=0.2.7=pypi_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.0.8=py_0
ncurses=6.2=he1b5a44_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
notebook=6.3.0=pypi_0
numpy=1.22.2=pypi_0
nvidia-ml-py3=7.352.0=pypi_0
olefile=0.46=py_0
opencensus=0.7.11=pypi_0
opencensus-context=0.1.2=pypi_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.0.5=pypi_0
openssl=1.1.1m=h7f8727e_0
osqp=0.6.1=py38hcb8c335_2
packaging=20.4=py_0
pandas=1.1.3=py38he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.7.0=py_0
patchelf=0.13=h295c915_0
pathtools=0.1.2=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pillow=9.0.1=pypi_0
pip=20.2.4=py38_0
postgresql=12.3=h8573dbc_0
prometheus-client=0.8.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0=pypi_0
psutil=5.7.3=pypi_0
psycopg2=2.8.5=py38h1ba5d50_0
ptyprocess=0.6.0=py38_0
py-spy=0.3.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.6.0=py_0
pycparser=2.20=pypi_0
pyflakes=2.2.0=py_0
pyglet=1.5.0=pypi_0
pygments=2.7.1=py_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.10.2=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2020.1=py_0
pyvirtualdisplay=2.1=py38h578d9bd_1
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py38he6710b0_1
qt=5.9.7=h5867ecd_1
ray=1.0.0=pypi_0
readline=8.0=he28a2e2_2
redis=3.4.1=pypi_0
requests=2.24.0=pypi_0
rsa=4.6=pypi_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
scs=2.1.2=py38hfe171b0_2
seaborn=0.11.0=py_0
send2trash=1.5.0=pypi_0
sentry-sdk=1.5.5=pypi_0
setuptools=50.3.0=py38hb0f4dca_1
shortuuid=1.0.8=pypi_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py_0
smmap=5.0.0=pypi_0
sniffio=1.2.0=pypi_0
soupsieve=2.0.1=pypi_0
sqlite=3.33.0=h62c20be_0
suitesparse=5.7.2=h717dc36_0
tbb=2020.2=hc9558a2_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=pypi_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hed695b0_0
tomlkit=0.7.0=pypi_0
torchaudio=0.8.1=pypi_0
torchvision=0.9.1+cu111=pypi_0
tornado=6.0.4=py38h7b6447c_1
tqdm=4.52.0=pyhd3deb0d_0
traitlets=5.0.5=py_0
typing-extensions=3.7.4.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzcode=2020a=h516909a_0
urllib3=1.25.11=pypi_0
wandb=0.12.10=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
wheel=0.35.1=py_0
xz=5.2.5=h516909a_1
yarl=1.6.2=pypi_0
yaspin=2.1.0=pypi_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h516909a_1009
zstd=1.4.5=h9ceee32_0
